reorganize the cli codes
